feat: add StepsHeader custom component
Adds a new StepsHeader custom component. The Steps component uses
h3 tags for each of the list items which causes the headers to
show up in the sidebar. We don't want these items to show up in
the sidebar. Comments of an issue in the Nextra GitHub suggest
using something that looks like `{</h3>My Header</h3>}` but this
is clumsy and confusing. Using a custom tag is much nicer.

Also has the side-effect of setting ourselves up for using other
custom tags in the future.