feat: add gfm and table linting
Updates linting rules so that we are properly linting for GitHub
Flavored Markdown (this is what Nextra uses) and so that we also
properly lint Markdown tables. Note that this will escape certain URL
characters like ampersands. This is intentional. An additional URL
remapping rule has been added to the Lychee config so that the escaped
URLs don't break link checking.