Add the year for Nov 11th update
When we put the reason for fault proofs being unavailable as "because of the EVM update" back in 2022 it was obvious that the update was in 2021. Now it's not so clear.

So I think either add the year, or provide a better reason for cannon fault proofs not being available yet.

(I happened to see this while updating links on Lattice's redstone.xyz).