Add ports for L2 and L1 dev nodes
Here is the command output of `make devnet-up` for reference:
```
[+] Running 3/3
 ✔ Container ops-bedrock-l1-1     Started                                                                                                                                                                                                                                0.1s 
 ✔ Container ops-bedrock-l1-bn-1  Started                                                                                                                                                                                                                                0.2s 
 ✔ Container ops-bedrock-l1-vc-1  Started                                                                                                                                                                                                                                0.4s 
[INFO|10-12-2024 08:43:12] Trying 127.0.0.1:8545
[INFO|10-12-2024 08:43:12] Connected 127.0.0.1:8545
[INFO|10-12-2024 08:43:12] Waiting for RPC server at 127.0.0.1:8545
[INFO|10-12-2024 08:43:12] RPC server at 127.0.0.1:8545 ready
[INFO|10-12-2024 08:43:12] L2 genesis and rollup configs already generated.
[INFO|10-12-2024 08:43:12] Bringing up L2.
WARN[0000] The "DG_TYPE" variable is not set. Defaulting to a blank string. 
WARN[0000] The "L2OO_ADDRESS" variable is not set. Defaulting to a blank string. 
WARN[0000] The "DGF_ADDRESS" variable is not set. Defaulting to a blank string. 
WARN[0000] The "PROPOSAL_INTERVAL" variable is not set. Defaulting to a blank string. 
WARN[0000] The "ALTDA_ENABLED" variable is not set. Defaulting to a blank string. 
WARN[0000] The "ALTDA_SERVICE" variable is not set. Defaulting to a blank string. 
WARN[0000] The "DA_TYPE" variable is not set. Defaulting to a blank string. 
WARN[0000] The "DGF_ADDRESS" variable is not set. Defaulting to a blank string. 
WARN[0000] The "ALTDA_GENERIC_DA" variable is not set. Defaulting to a blank string. 
WARN[0000] The "ALTDA_ENABLED" variable is not set. Defaulting to a blank string. 
WARN[0000] The "ALTDA_SERVICE" variable is not set. Defaulting to a blank string. 
[+] Running 1/1
 ✔ Container ops-bedrock-l2-1  Started                                                                                                                                                                                                                                   0.1s 
[INFO|10-12-2024 08:43:12] Trying 127.0.0.1:9545
[INFO|10-12-2024 08:43:12] Connected 127.0.0.1:9545
[INFO|10-12-2024 08:43:12] Waiting for RPC server at 127.0.0.1:9545
[INFO|10-12-2024 08:43:12] Waiting for RPC server at 127.0.0.1:9545
[INFO|10-12-2024 08:43:13] RPC server at 127.0.0.1:9545 ready
[INFO|10-12-2024 08:43:13] Using batch inbox 0x00289c189bee4e70334629f04cd5ed602b6600eb
[INFO|10-12-2024 08:43:13] Using DisputeGameFactory 0xeCb92a686D1ab066fc4E559A305FEB75DD512377
[INFO|10-12-2024 08:43:13] Bringing up `op-node`, `op-proposer` and `op-batcher`.
WARN[0000] The "L2OO_ADDRESS" variable is not set. Defaulting to a blank string. 
```